#134 implemented configEventchannel

+++ handlers/config.js
@@ -249,19 +249,23 @@ async function confirmGuildConfig(guild) {
 }
 
 /**
- *
+ * send all events to this channel
  * @param {Message} msg
+ * @param {Array} msgParms
  * @param {GuildModel} guildConfig
  */
-async function handleConfigEventChannel(msg, guildConfig) {
+async function handleConfigEventChannel(msg, msgParms, guildConfig) {
     try {
         if (await users.hasRoleOrIsAdmin(msg.member, guildConfig.arole)) {
-            let stringParam = msg.content.substring((guildConfig.prefix + 'config eventchannel').length + 1);
+            let channelTest;
+            let stringParam = msgParms.length > 0 ? msgParms[0].value : '';
             if (stringParam.trim() == '') {
                 guildConfig.channelForEvents = undefined;
             } else {
-                stringParam = stringParam.substring(2, stringParam.length - 1);
-                let channelTest = await msg.guild.channels.resolve(stringParam);
+                if (stringParam.startsWith('<')) {
+                    stringParam = stringParam.substring(2, stringParam.length - 1);
+                }
+                channelTest = await msg.guild.channels.resolve(stringParam);
                 if (!channelTest) {
                     throw new Error(`Could not locate channel: ${stringParam}`);
                 }
@@ -269,8 +273,10 @@ async function handleConfigEventChannel(msg, guildConfig) {
             }
             await guildConfig.save();
             GuildCache[msg.guild.id] = guildConfig;
-            await utils.sendDirectOrFallbackToChannel({ name: 'Config Event Channel', value: `Event Channel now set to: \`${guildConfig.channelForEvents}\`.` }, msg);
-            await msg.delete();
+            await utils.sendDirectOrFallbackToChannel({ name: 'Config Event Channel', value: `Event Channel now set to: \`${guildConfig.channelForEvents ? channelTest.name : guildConfig.channelForEvents}\`.` }, msg);
+            if (msg.deletable) {
+                await msg.delete();
+            }
         } else {
             throw new Error(`please ask an \`approver role\` to configure.`);
         }
